WP-204 functionality migrated from slide to slides component
